Migrated the static filter functions to the FilterService
